Environment deletion
Adds environment deletion through editor and invokes new event. Adds new envType parameter to environmental tiles

+++ Assets/Scripts/HexGrid/BuildingManager.cs
@@ -60,8 +60,10 @@ public class BuildingManager : MonoBehaviour
     public static UnityEvent DisableEditing = new UnityEvent();
 
     //events to mark building interactions
-    public static UnityEvent<Building, Vector3Int> BuildingClicked = new UnityEvent<Building, Vector3Int>();
     public static UnityEvent<EnvironmentTile, Vector3Int> EnvironmentClicked = new UnityEvent<EnvironmentTile, Vector3Int>();
+    public static UnityEvent<EnvironmentTile, Vector3Int> EnvironmentDeleted = new UnityEvent<EnvironmentTile, Vector3Int>();
+
+    public static UnityEvent<Building, Vector3Int> BuildingClicked = new UnityEvent<Building, Vector3Int>();
     public static UnityEvent<Building, Vector3Int> BuildingPlaced = new UnityEvent<Building, Vector3Int>();
     public static UnityEvent<Building, Vector3Int> BuildingDeleted = new UnityEvent<Building, Vector3Int>();
 
@@ -247,7 +249,10 @@ public class BuildingManager : MonoBehaviour
                 PlaceBuilding(offsetCoord, activeStructure);
                 break;
             case EditMode.Delete:
-                DeleteBuilding(offsetCoord);
+                if (!DestroyEnvironment(offsetCoord))
+                {
+                    DeleteBuilding(offsetCoord);
+                }
                 break;
         }
     }
@@ -392,6 +397,22 @@ public class BuildingManager : MonoBehaviour
         BuildingDeleted.Invoke(building, offsetCoord);
 
         return true;
+    }
+
+
+
+    //attempts to destroy an environmental tile in a given location
+    public bool DestroyEnvironment(Vector3Int offsetCoord)
+    {
+        if (IsWithingRange(offsetCoord) && objectMap.GetTile(offsetCoord) is EnvironmentTile envTile)
+        {
+            objectMap.SetTile(offsetCoord, null);
+            EnvironmentDeleted.Invoke(envTile, offsetCoord);
+            return true;
+        }
+
+        FailedDestroy.Invoke();
+        return false;
     }
